+++ src/components/shapes/fractions/index.js
@@ -7,5 +7,6 @@ export { default as SquareQuarterSquareFraction } from "./SquareQuarterSquareFra
 export { default as SquareCrossFraction } from "./SquareCrossFraction";
 export { default as SquareIsoscelesTriangleFraction } from "./SquareIsoscelesTriangleFraction";
 export { default as SquareRectangleThin8thFraction } from "./SquareRectangleThin8thFraction";
+export { default as SquareHalfRectangle8thFraction } from "./SquareHalfRectangle8thFraction";
 export { default as RectangleFraction } from "./RectangleFraction";
 export { default as HouseFraction } from "./HouseFraction";